+++ src/sbr/mod.rs
@@ -8,4 +8,6 @@ pub mod transport;
 
 pub use i2c::*;
 pub use parse::*;
-pub use transport::{IstwiSbrTransport, SbrTransport, SbrTransportError, VfioI2cSbrTransport};
+pub use transport::{
+    Bar1MmapSbrTransport, IstwiSbrTransport, SbrTransport, SbrTransportError, VfioI2cSbrTransport,
+};